+++ src/test/java/school/redrover/ArMobiTest.java
@@ -1,10 +1,13 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.testng.Assert;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 public class ArMobiTest {
@@ -13,10 +16,21 @@ public class ArMobiTest {
 
     private final By getErrorText = By.xpath("//div[@style='text-align: center; margin-bottom: 20px; color: rgb(255, 0, 0);']");
 
+    WebDriver driver = new ChromeDriver();
+
+    @BeforeTest
+    public void initWebDriver() {
+        driver = new ChromeDriver();
+        driver.manage().window().setSize(new Dimension(1820,1080));
+    }
+
+    @AfterTest
+    public void closeWebDriver() {
+        driver.quit();
+    }
     @Test
     public void removePaswordTest() throws InterruptedException {
 
-        WebDriver driver = new ChromeDriver();
         driver.get("http://23.105.246.172:5000/login");
 
         Thread.sleep(1000);
@@ -31,16 +45,13 @@ public class ArMobiTest {
 
         Assert.assertEquals(getError,"Неправильный логин или пароль");
 
-        driver.quit();
     }
 
     @Test
     public void regaTest() throws InterruptedException {
 
-        WebDriver driver = new ChromeDriver();
-
         driver.get("http://23.105.246.172:5000/login");
-        Thread.sleep(1000L);
+        Thread.sleep(1000);
 
         driver.findElement(By.xpath("//h2[@class='ant-typography h2_m Login__restore-text']")).click();
 
@@ -48,12 +59,11 @@ public class ArMobiTest {
         inputMail.sendKeys("n-k-65@list.ru");
 
         driver.findElement(By.xpath("//button[@class='ant-btn ant-btn-default authButton big colorPrimary ']")).click();
-        Thread.sleep(1000L);
+        Thread.sleep(1000);
 
         String getPasError = driver.findElement(getPaswordText).getText();
 
         Assert.assertEquals(getPasError,"Мы отправили по адресу n-k-65@list.ru ссылку для восстановления доступа");
 
-        driver.quit();
     }
 }
